Okay It's Working Now

+++ Airplane Simulator/Airplane Simulator/Simulator.h
@@ -109,7 +109,7 @@ public:
 		// FIXME: Calculate and display the average wait time for the landing queue
 		double landing_queue_average_wait_time;
 		if (landing_queue->get_num_served() > 0)
-			landing_queue_average_wait_time = landing_queue->get_total_wait() / landing_queue->get_num_served();
+			landing_queue_average_wait_time = ((double)landing_queue->get_total_wait() / landing_queue->get_num_served());
 		else
 			landing_queue_average_wait_time = 0;
 		std::cout << "Average wait time for the arrival queue: " << landing_queue_average_wait_time << std::endl;
@@ -122,7 +122,7 @@ public:
 		// FIXME: Calculate and display the average wait time for the departure queue
 		double departure_queue_average_wait_time;
 		if (departure_queue->get_num_served() > 0)
-			departure_queue_average_wait_time = departure_queue->get_total_wait() / departure_queue->get_num_served();
+			departure_queue_average_wait_time = ((double)departure_queue->get_total_wait() / departure_queue->get_num_served());
 		else
 			departure_queue_average_wait_time = 0;
 		std::cout << "Average wait time for the departure queue: " << departure_queue_average_wait_time << std::endl;
